Add app skeleton and styling
The app skeleton has been largely created, i.e. the page listing
root goals and the page for specific root goals, listing subgoals.

Some basic navigation is implemented, but this should be replaced
with something such as Redux in future.

Styling is applied using Sass, normalize.css, and fontawesome
(which are already in the package.json).