Begin implementation of subgoal list editing
The following two core features are partially implemented:
- Subgoal drag & drop (via react-beautiful-dnd)
- New subgoal indentation/depth modification (to add subgoals
  to different levels in the hierarchy)

The front-end mechanics for these features are largely in place,
but the 'back-end' modification of the actual goal list is not
yet implemented.

Logic functions for flattening and searching the goal tree have
been added (to src/logic). These are utilized to generate indices
from the goal tree, such that it is possible to rearrange it
(for drag and drop).

Due to constraints in the immutability of Redux state, it is now
preferable to pass goal IDs rather than instances, so that
inconsistencies do not occur.