Change setSubgoalTitle to setGoalName
This naming makes matters simpler: if a goal is being considered, then
it may be a goal or a subgoal. Further, name corresponds better to the
internal naming of the goal's title property. Finally, we also remove
the goal object itself from the parameters of the action so that we are
not passing around outdated references, and instead use its ID.