Only allow marking goals without children as complete
There is now a distinction to how a goal can be classed as complete:
either it is *explicitly* complete, which only leaf nodes are permitted
to be (and the user has to mark it as complete in this case), or it is
*implicitly* complete, which a goal is classed as if all of its children
are complete.

The problem this solves is an inconsistency between the completeness of
goals. Goals must at all times satisfy the following criteria for being
complete:
- If a goal is marked as complete, it is complete
- If a goal's parent is complete, then it is also complete
- If a goal is complete, all of its children are also complete

The rationale behind this is that it does not make sense to have a
parent that is complete when its children are not, and nor does the
reverse make sense.

This commit is one of many options to fulfill these criteria, which I
opted for as it was the simplest option. The other options were as
follows:

1. *Automatic marking*
   When a goal is marked as complete, its parent and children are
   updated so as to fit the criteria listed above. In this scenario,
   there is only one 'completed' field.

   The issue with this option is that it'd also involve implementing
   sensible logic for what happens to the children of a goal when its
   parent is marked as incomplete: if the children were independently
   marked as complete, they should stay; if the children were marked
   as complete to comply with the criteria, should they then all be
   marked as incomplete?

2. *Implicit completeness*
   In this scenario, goals are never explicitly modified to comply with
   the criteria. Instead, they will simply appear as if they are
   complete if it is necessary to comply with the criteria.

   The issues with this option are that there would be no ability to
   have a record of when/how/why a goal became 'complete', and that it
   would be confusing as to how marking a child goal affects the parent,
   if the parent is already complete (it wouldn't make a difference
   whether the child goal was set as complete/incomplete by the user,
   because it would be overridden by the parent's completeness, which
   might be quite hard to figure out).

To avoid these issues and to simplify the process of marking goals as
complete to the maximal amount while still abiding by the criteria, we
adopt the option that disallows marking any inodes (i.e. goals with
children) as either complete or incomplete. They will appear as
completed, implicitly, if all of their children are completed (either
explicitly or implicitly). This way, it is easy to identify why, where,
and how a goal was completed, and there is clear behaviour for what
happens if you try and mark a parent (i.e. you can't in the first
place).